Add The Foamiest Fix, some config tweaks
- Set dormantChunkCache to 32
- Enable alwaysSetupTerrainOffThread, as recommended by FoamFix
- Enable replaceVanillaBucketModel, as much as I hate it
- Enable removeErroring(Tile)Entities
- Disable baby zombies (:D :D :D)
- Disable the Forge light pipeline

+++ prepare.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 find ./ -name '*.url.txt' | while read file; do
 	basename=${file%.url.txt}
-	url=`cat $file`
-	echo Downloading $url...
-	curl -s "$url" -o "$basename"
+	if [ ! -e "$basename" ]; then
+		url=`cat $file`
+		echo Downloading $url...
+		curl -s "$url" -o "$basename"
+	else
+		echo Skipping download of existing file `basename $basename`
+	fi
 done
